Implement the get-function function.
Add a Function(name, Option<SassFunction>) alternative to css::Value,
and modify the call function to take such a Value.  The warning when
the Value is a Literal is more motivated now.

To do this, I had to derive Ord for SassFunction, which required
implementing it manually for FuncImpl and deriving it for Selectors,
FormalArgs and sass::Item.