Report file position of parse errors
As requested in #46.

This is just a start.  It reports where a parse error occurd, but due
to the nature of the parser, the error "occurs" at the start of any
top-level block containing the error.  It should be possible to modify
the parser so that more specific parse errors are found.