Improve error handling in functions
Remove the catch-all code that retained the function call if there was
an error in executing the function.  Instead, report such errors, but
do support "bad" arguments to some (color handling) functions that
also are css builtin functions.

This partially solves the problem described in #41; now the example
terminate promptly with an error message, as the special case of
calling string-handling functions with an url(...) parameter is not
yet supported.