Isolate fs Path handling.
Paths in sass are url paths, but the path to the initial file
to open from command-line or api is a std::fs::Path.  On unix
this does not really matter, but on other operating systems,
like windows, there might be important differences.

So isolate the std::fs::Path handling to where it is needed and
try to keep the url-path handling clean in other places.