Replace the span type
Use my own Span type instead of the locatedspan crate, and make
SourcePos behave like a Span over an owned (reference counted)
SourceFile.  My own span is the same, but borrows the SourceFile so it
can implemnet Copy, which is kind of required by nom.

Should give fewer allocations as the number of required positions to
keep track of grows.