Be more flexible in custom values.
The values of custom properties (css variables) are not parsed as
ordinary values, but as a kind of extended strings.  Up to now, it was
either a quoted string or a unquoted specially extended string.  Now,
it is instead the extended string directly, but the extended string
can contain quoted strings.

Fixes #175.