Use logical selectors in nest.
This adds supports for sass backref (`&`) is `LogicalSelectorSet` and
introduces a new type, `CssSelectorSet`, which is guaranteed to not
contain backrefs.

Functions that are intended to not accept backrefs are moved to
`CssSelectorSet`.  The `nest` function is also created as a
`CssSelectorSet` method, as the first argument may not contain
backrefs.  The other argument is a `LogicalSelectorSet`, though.